+++ server/Domain/Services/Impl/ObjectiveService.ts
@@ -2,6 +2,7 @@ import { injectable, inject } from 'inversify';
 import { IObjectiveService } from '../IObjectiveService';
 import ObjectiveEntity from '../../../Data/Entities/ObjectiveEntity';
 import CareerDayEntity from '../../../Data/Entities/CareerDayEntity';
+import ApplicationError from './ApplicationError';
 
 import {
   IObjectiveRepository,
@@ -29,9 +30,18 @@ export class ObjectiveService implements IObjectiveService {
   }
 
   public async addObjective(data: any): Promise<ObjectiveEntity> {
-    const objective = new ObjectiveEntity(data);
+    const careerDay = await this._careerDayRepository.findById(data.CareerDayId);
 
-    return this._objectiveRepository.create(objective);
+    if (!careerDay.Archived) {
+      if (data.EmployeeExternalId === careerDay.EmployeeExternalId
+        && data.UnitManagerExternalId === careerDay.UnitManagerExternalId) {
+        const objective = new ObjectiveEntity(data);
+
+        return this._objectiveRepository.create(objective);
+      }
+      throw (new ApplicationError('You are  not manager of this employee.', 403));
+    }
+    throw (new ApplicationError('No one can add objective in which Career Day was archived.', 403));
   }
 
   public async updateObjective(id: number, title: string, description: string) {
@@ -43,15 +53,17 @@ export class ObjectiveService implements IObjectiveService {
       objective.Description = description;
 
       return this._objectiveRepository.update(objective);
-    } else {
-      throw {
-        code: 403,
-        massage: 'No one can edit objective in which Career Day was archived .',
-      };
     }
+    throw (new ApplicationError('No one can edit objective in which Career Day was archived.', 403));
   }
 
   public async deleteObjective(id: number): Promise<void> {
-    await this._objectiveRepository.remove({ where: { id } });
+    const careerDay = await this._careerDayRepository.findById(id);
+
+    if (!careerDay.Archived) {
+      await this._objectiveRepository.remove({ where: { id } });
+    } else {
+      throw (new ApplicationError('No one can delete objective in which Career Day was archived.', 403));
+    }
   }
 }
